FAKE NEWS/requirements.txt
torch
torchvision
transformers
pandas
numpy
scikit-learn
streamlit
Pillow
defusedxml
